[CAMERAWG] - Mostly under-the-hood upgrades to the niawg control code, in preparation for attempting to implement streaming for rearrangement stuffs. - Implemented niawgPair template typedef which drastically reduced the number of lines & imo improved readability of niawg control code. - More throwers, less int returns - introduced str() function - Introduced ScriptStream class imported from master - some other small changes.

+++ Camerawg/VariableSystem.cpp
@@ -194,16 +194,17 @@ bool VariableSystem::deleteVariable()
 	return false;
 }
 
+
 variable VariableSystem::getVariableInfo(int varNumber)
 {
 	return currentVariables[varNumber];
 }
 
-unsigned int VariableSystem::getCurrentNumberOfVariables()
+
+unsigned long long VariableSystem::getCurrentNumberOfVariables()
 {
 	return currentVariables.size();
 }
-
 
 
 bool VariableSystem::clearVariables()
